+++ api/config/__init__.py
@@ -8,15 +8,17 @@ This package provides integration with the GitHub API, including:
 - Repository metadata management
 """
 
-from github.github_api import GitHubAPI
-from github.cache_manager import CacheManager, cache_manager
-from github.github_repo_manager import GitHubRepoManager
-from github.fingerprint_manager import FingerprintManager
+from config.github_api import GitHubAPI
+from config.cache_manager import CacheManager, cache_manager
+from config.github_repo_manager import GitHubRepoManager
+from config.fingerprint_manager import FingerprintManager
+from config.fine_tuning import SemanticModel
 
 __all__ = [
     'GitHubAPI',
     'GitHubRepoManager',
     'CacheManager',
     'cache_manager',
-    'FingerprintManager'
+    'FingerprintManager',
+    'SemanticModel'
 ]
