+++ api/config/fingerprint_manager.py
