+++ api/host.json
@@ -10,6 +10,6 @@
   },
   "extensionBundle": {
     "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[4.12]"
+    "version": "[4.*, 5.0.0)"
   }
 }
